fix: close 24 second-pass audit bugs
Second-pass deep audit of crates previously reported clean. Real
bugs + one fabricated round-1 fix now actually applied.

High (panic/DoS):
- html linkpreview: <link> w/o attrs sliced [6..5] -> panic
- html quote scanner: now matches opener, not either quote
  (round-1 builder fabricated this fix; never applied)

Med (resource/security):
- fts5: quote terms so AND/OR/NOT/NEAR are literals, not syntax
- fts5: term length uses char count, not UTF-8 bytes (CJK/emoji)
- i2p: 30s read timeout frees inbound slot on stalling peer
- freenet: 5s timeout on block_on — hung node can't starve poll
- waveform: cap container decode at 64 MiB samples (voice.rs mirror)
- cas: guard from>to slice panic on short/corrupt chunk file
- telemetry: wrap sets len=cap so read_all sees pre-wrap history
- telemetry: header written atomically (single slice), no torn reads
- telemetry: reject capacity shrink — was silently wiping data

Low (correctness/defense):
- custom_profile: include index in default_node id (dup same-ms)
- entities: drop phantom semicolon on bare "&#"
- url: block RFC 2544 198.18.0.0/15 (SSRF depth)
- format_duration: round, not truncate (59.9s -> 1:00)
- thread_governor: probe available_parallelism instead of hardcoded 8
- normalized_store: notify StoreCleared before clear()
- ranking: u64 engagement sum (u32 overflow)
- thumbhash: checked dims arithmetic (32-bit overflow)
- format_count: +Inf clamps to "0"
- relay: assert envelope str<=128; dedup payload_digest

+++ feed-core/src/ranking.rs
@@ -40,8 +40,10 @@ pub fn score_post_with_set(
     now_secs: f64,
 ) -> f64 {
     let hours_ago = (now_secs - stats.created_at_secs).max(0.1) / HOUR_SEC;
-    let engagement =
-        stats.likes_count + stats.reposts_count + stats.zaps_count + stats.replies_count;
+    let engagement = (stats.likes_count as u64)
+        + (stats.reposts_count as u64)
+        + (stats.zaps_count as u64)
+        + (stats.replies_count as u64);
     let velocity = engagement as f64 / hours_ago;
     let wot_boost = (2.0 - stats.wot_distance.min(2) as f64).max(0.5);
     let recency_factor = 1.0 / (hours_ago + 1.0).log2();
